requirements.txt
# Requirements for EEG Motor Imagery Pipeline
mne
numpy
scikit-learn
torch
onnx
onnxruntime
pyyaml
pytest
